refactor(schedule): extract the scheduler loop + timer cell (phase 4)
Moves the production scheduler wiring out of index.ts into
lib/scheduler-loop.ts: the `schedulerTimer` module cell (now in exactly one
module), the settings-derived tick/catch-up accessors (ScheduleSettings +
DEFAULT/MIN constants + tickIntervalMs/catchUpWindowMs), the schedulerDeps
factory (bound to the live ledger + fire engine), the idempotent
runScheduleCutover, and startScheduler/stopScheduler with the verbatim
self-rescheduling cycle (resume-pauses → tick → heal).

CRITICAL ordering preserved: the activate() startup sequence stays in index
UNCHANGED — cutover → … → resumeDuePauses → runStartupCatchUp(schedulerDeps(),
catchUpWindowMs()) → startScheduler — and onShutdown still calls stopScheduler.
Only the functions moved; index imports them back. index sheds the ledger
claim/getCronFire, runSchedulerTick/DEFAULT_TICK_WINDOW_MS/SchedulerDeps,
migrateBakinSchedulesOffOpenClawCron, and runClaimedFire imports (now all
scheduler-loop-internal). 1,153 -> 1,060.

Pure behavior-preserving relocation — functions byte-identical, the cycle and
the activate ordering unchanged.

Verified: typecheck + eslint + full schedule suite (280/0) + full suite
(5124/0) + binary build + DOCKERIZED-RIG E2E against real OpenClaw (isolated):
an every-minute schedule fired AUTOMATICALLY via the scheduler tick —
occurrence runId `sch_…:2026-06-23T15:31:00.000Z` (not manual-), status
success, task created — proving startScheduler → cycle → runSchedulerTick →
schedulerDeps.fire → runClaimedFire end to end on the live stack.
